DTGB-871: Fix GrumpPHP error when installing in project
There is an error reported during the installation of the base theme.
The error is caused by the install scripts that triggers an composer
install in the base theme directory.

The post-install scripts in the composer.json file want to init GrumPHP
but it is not available since dev packages are not installed.

This post install step is not necessary when installing the Base theme
for development; GrumPHP inits itself the moment it is installed.

See https://github.com/phpro/grumphp/blob/master/README.md:

> When the package is installed, GrumPHP will attach itself to the git
> hooks of your project. You will see following message in the composer
> logs:
>
> Watch out! GrumPHP is sniffing your commits!