HEAT-6508209: Fix display of file "table"
I honestly have no idea how this has ever worked. We recently updated
our environment to PHP8 and that's about when this problem started.
However, I don't see (looking at the code) how PHP7 would display this
correctly. The theme_item_link function clearly states in [its
documentation] that the items parameter should be "An array of items
to be displayed in the list. If an item is a string, then it is used
as is. If an item is an array, then the "data" element of the array is
used as the contents of the list item. If an item is an array with a
"children" element, those children are displayed in a nested list. All
other elements are treated as attributes of the list item element." This
function hasn't changed in _9 years_ in Drupal core.

The code calling it in this theme hasn't changed in _6 years_. However,
before this fix, we called it with an array of arrays (meaning each item
was an array and _should have_ had a `data` key, but it didn't. So now
(presumably since the PHP8 upgrade, but again I have no clue why it
would work in PHP7), since we didn't specify a key, the key was
effectively `0`. And since "All other elements are treated as attributes
of the list item element.", we ended up with empty list items that had
an attribute `0` with the link to the file as the value, and not the
link to the file in the "body" of the list item.

[its documentation]:
https://api.drupal.org/api/drupal/includes%21theme.inc/function/theme_item_list/7.x