HEAT-6798135: Fix capitalization vesta mail notes
Twig's `capitalize` filter makes the first letter uppercase, but _all_
other letters lowercase. This was probably added for single-word labels
like "e-mail" or "contact", to make it consistent. But in some cases the
key users use entire sentences, and this filter makes the letters after
punctiation lowercase, which isn't right. This commit replaces the
`capitalize` filter with Twig's equivalent of `ucfirst`. This way we
keep backward compatibility for single-word labels, but we also make
sure lables with entire sentences are displayed correctly.