HEAT-6832177: Fix cache key for teaser_with_links
When rendering field_contacts, vesta entities are rendered in different
view modes depending on how many entities are referenced in the field.
We were just altering the render array, and handling the
CacheableMetadata, but CacheableMetadata does not handle cache keys,
only cache contexts, tags and max-age. To be safe, we just rebuild the
entire render array in the new view mode, and still handle the
CacheableMetadata like before. This ensures we also have the correct
cache keys in the render arrays.